requirements.txt
python-dotenv>=1.0.0
scenedetect[opencv]>=0.6.4
opencv-python>=4.9.0
Pillow>=10.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
transformers>=4.41.0,<5.0.0
torch>=2.3.0
sentencepiece>=0.2.0
timm>=0.9.0
einops>=0.7.0
tokenizers>=0.15.0,<0.20.0
tqdm>=4.66.0
numpy>=1.26.0
